Improved Now Streaming and Coming Soon
The "Now Streaming" and "Coming Soon" categories were pulling 1 page worth of results and then filtering down to ensure that the movies were actually correctly categorized. This left the lists too small. I increased the number of pages to 5 for both the refresh_movies function as well as the startup behavior in the __main__ block. Initially, some of this logic had only been applied to the refresh_movies function, meaning at startup, the list was too short.